Calorie-Expenditure-Prediction/pyproject.toml
[project]
name = "calories_prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
]
